+++ auxiliary/jacobian.cpp
@@ -116,26 +116,26 @@ void Jacobian::calc2(double * Ans, const double * u, const double * v)
 	//generate_right_part(&rp[0], m_, (right_part_cb_t)diff_2_cos_rp, (void*)u);
 	diff2_cos_.mult_vector(&tmp[0], &u_in[0]);
 	diff2_cos_rp_.mult_vector(&rp[0], &u_in_bnd[0]);
-	vec_sum(&rp[0], &rp[0], &tmp[0], rp.size());
+	vec_sum(&rp[0], &rp[0], &tmp[0], (int)rp.size());
 	idt_.solve(&pt1[0], &rp[0]);
 
 	//generate_right_part(&rp[0], m_, (right_part_cb_t)diff_1_rp, (void*)v);
 	diff1_.mult_vector(&tmp[0], &v_in[0]);
 	diff1_rp_.mult_vector(&rp[0], &v_in_bnd[0]);
-	vec_sum(&rp[0], &rp[0], &tmp[0], rp.size());
+	vec_sum(&rp[0], &rp[0], &tmp[0], (int)rp.size());
 	idt_.solve(&tmp[0], &rp[0]);
 	vec_mult(&pt1[0], &pt1[0], &tmp[0], (int)pt1.size());
 
 	//generate_right_part(&rp[0], m_, (right_part_cb_t)diff_1_cos_rp, (void*)u);
 	diff1_cos_.mult_vector(&tmp[0], &u_in[0]);
 	diff1_cos_rp_.mult_vector(&rp[0], &u_in_bnd[0]);
-	vec_sum(&rp[0], &rp[0], &tmp[0], rp.size());
+	vec_sum(&rp[0], &rp[0], &tmp[0], (int)rp.size());
 	idt_.solve(&pt2[0], &rp[0]);
 
 	//generate_right_part(&rp[0], m_, (right_part_cb_t)diff_2_rp, (void*)v);
 	diff2_.mult_vector(&tmp[0], &v_in[0]);
 	diff2_rp_.mult_vector(&rp[0], &v_in_bnd[0]);
-	vec_sum(&rp[0], &rp[0], &tmp[0], rp.size());
+	vec_sum(&rp[0], &rp[0], &tmp[0],(int) rp.size());
 	idt_.solve(&tmp[0], &rp[0]);
 	vec_mult(&pt2[0], &pt2[0], &tmp[0], (int)pt1.size());
 
@@ -146,7 +146,7 @@ void Jacobian::calc2(double * Ans, const double * u, const double * v)
 void Jacobian::calc2t(double * Ans, const double * u, const double * v)
 {
 	calc2(Ans, u, v);
-	vec_mult_scalar(Ans, Ans, -1.0, m_.inner.size());
+	vec_mult_scalar(Ans, Ans, -1.0, (int)m_.inner.size());
 }
 
 /**
@@ -154,14 +154,14 @@ void Jacobian::calc2t(double * Ans, const double * u, const double * v)
  */
 Jacobian::Jacobian(const Mesh & m): m_(m), 
 	idt_((int)m.inner.size()),
-	diff1_(m.inner.size()),
-	diff2_(m.inner.size()),
-	diff1_cos_(m.inner.size()),
-	diff2_cos_(m.inner.size()),
-	diff1_rp_(m.inner.size()),
-	diff2_rp_(m.inner.size()),
-	diff1_cos_rp_(m.inner.size()),
-	diff2_cos_rp_(m.inner.size())
+	diff1_((int)m.inner.size()),
+	diff2_((int)m.inner.size()),
+	diff1_cos_((int)m.inner.size()),
+	diff2_cos_((int)m.inner.size()),
+	diff1_rp_((int)m.inner.size()),
+	diff2_rp_((int)m.inner.size()),
+	diff1_cos_rp_((int)m.inner.size()),
+	diff2_cos_rp_((int)m.inner.size())
 {
 	generate_matrix(idt_, m, id_cb, 0);
 
@@ -186,6 +186,6 @@ void Jacobian::calc1(double * Ans, const double * u, const double * v, const dou
 void Jacobian::calc1t(double * Ans, const double * u, const double * v, const double * bnd)
 {
 	calc1(Ans, u, v, bnd);
-	vec_mult_scalar(Ans, Ans, -1.0, m_.ps.size());
+	vec_mult_scalar(Ans, Ans, -1.0, (int)m_.ps.size());
 }
